Applied (unofficial) Patch 0.9.0 => 0.10.0
git-svn-id: https://vistrpndsvn1.vi.vector.int/svn/CANoe_Components/OpenRTI/trunk/src/OpenRTI@461100 fc9017fd-cfa4-6c48-8536-0279c867a896

+++ posix/ClockPosix.h
@@ -45,7 +45,7 @@ struct OPENRTI_LOCAL ClockPosix {
   static struct timespec toTimespec(const uint64_t& nsec)
   {
     struct timespec ts;
-    ts.tv_nsec = nsec % 1000000000;
+    ts.tv_nsec = nsec % uint64_t(1000000000);
     uint64_t sec = nsec/uint64_t(1000000000);
     if (uint64_t(std::numeric_limits<time_t>::max()) < sec) {
       ts.tv_nsec = 999999999;
